Add files via upload
The algorithm inferring "true" allele age for all the 1000 markers of a given dataset. The example is given for the N3 dataset. The inputs are the file storing all true and inferred TMRCA from N3, and the file storing the inferred allele ages for the 1000 variants considered.
The outputs are 2 files, either comprising the 1000 "true" allele ages or the 1000 inferred allele ages, all population-scaled.